+++ src/components/map/status-indicator.tsx
